corte de control completado

+++ Programacion Estructurada Basica/includes/myShortCuts.h
@@ -2,10 +2,10 @@
 #define FLUSH_H
 #include <stdio.h>
 
-void flush(FILE *stream);
-void scanfInt(int *n, int li, int ls, int c);
-void scanfFloat(float *n, float li, float ls, int c);
+void flush(FILE* stream);
+void scanfInt(int* n, int li, int ls, int c);
+void scanfFloat(float* n, float li, float ls, float c);
 void fsgets(char s[], int n);
-void Err(char *msg);
+void Err(char* msg);
 
 #endif
